Support regex type
